python_code/requirement.txt
pymongo==4.0.2
dnspython==2.2.1
google-cloud-pubsub==2.11.0